Hack to avoid crashing with duplicate selected vars
This is a temporary hack to see if we can generate all factsheets for the website. The logic of these methods often assumes unique variable names, but that is not true in the var dictionary; this should be reworked and var_readable names should be updated to be unique and logic should be reworked.