Add in downloadACSPlacePopEst
From a separate call to get overall population estimates for the US at the place level. This variable isn't in any tables we're already requesting at the place level; it makes sense that we do this as a separate call. May be worth refactoring this at some point, but this is fine for now as we're getting the package restructured.