pwd_total_commute now handled for tract level
This variable isn't available at the tract level; therefore, we need special handling to make sure that we can mutate it if need be. Not sure if this is for sure the handling that I want, or if I want to make this column NA. For now, this will run.